Fix links to Wikidata entities
Instead of hard-coding the assumption that properties are in the
Property: namespace and everything else is in the main namespace,
install a map of item ID letters to prefixes and consult that. The map
defaults to the prefixes on Wikidata (items in main namespace,
properties in Property: namespace, lexemes in Lexeme: namespace), but
can be overridden with wiki-specific configuration, e. g. from
QuickStatements’ config.json file.

Links to forms and senses additionally require the transformation of the
form- or sense-specific part (the “F1” of “L1-S1”) into the fragment of
the URL that otherwise refers to the containing lexeme (forms and senses
have no pages of their own). Accordingly, the function is also renamed,
since it no longer just prefixes the entity ID with something.